component file structure
