+++ package.json
@@ -20,6 +20,8 @@
   "dependencies": {
     "react": "16.3.1",
     "react-native": "~0.55.2",
-    "react-native-firebase": "^4.1.0"
+    "react-native-firebase": "^4.1.0",
+    "react-native-material-ui": "^1.22.3",
+    "react-navigation": "^2.3.1"
   }
 }
