+++ src/components/Dashboard.js
@@ -1,0 +1,11 @@
+import React from 'react';
+import { StyleSheet, Text, View } from 'react-native';
+
+const Dashboard = props => {
+  return (
+    <View>
+    </View>
+  )
+}
+
+export default Dashboard
