+++ src/components/Home.js
@@ -1,0 +1,11 @@
+import React from 'react';
+import { StyleSheet, Text, View } from 'react-native';
+
+const Home = props => {
+  return (
+    <View>
+    </View>
+  )
+}
+
+export default Home
