+++ src/components/Login.js
@@ -1,0 +1,14 @@
+import React from 'react';
+import { StyleSheet, Text, View, Image } from 'react-native';
+
+const Login = props => {
+  return (
+    <View>
+      <Image source={{uri:'https://en.wikipedia.org/wiki/Smiley#/media/File:SNice.svg'}} style={{width: 50}}/>
+      <Text>Hi</Text>
+      <Text>Hey</Text>
+    </View>
+  )
+}
+
+export default Login
