+++ src/components/index.js
@@ -1,0 +1,5 @@
+//this file is the 'central export' for our components
+
+export { default as Login } from './Login'
+export { default as Home } from './Home'
+export { default as Dashboard } from './Dashboard'
